--- src/cmds/ping.ts
@@ -1,10 +1,0 @@
-import FluorineClient from '@classes/Client';
-import Embed from '@classes/Embed';
-import { Message } from 'discord.js';
-
-export async function run(client: FluorineClient, message: Message) {
-    const embed = new Embed(client, message.guild.preferredLocale)
-        .setTitle('Ping')
-        .addLocaleField({ name: 'PING', value: `${client.ws.ping}ms` });
-    message.reply({ embeds: [embed] });
-}
