+++ src/components/help.ts
@@ -7,9 +7,7 @@ export const authorOnly = true;
 
 export async function run(client: FluorineClient, interaction: SelectMenuInteraction) {
     const [category] = interaction.values;
-    const commands = client.applicationCommands.chatInput.filter(
-        (c: ChatInputCommand) => c.category === category && !c.dev
-    );
+    const commands = client.commands.chatInput.filter((c: ChatInputCommand) => c.category === category && !c.dev);
 
     const fields: EmbedFieldData[] = commands.map(c => ({
         name: `/${c.data.name_localizations[interaction.locale] ?? c.data.name}`,
